Re-organize migration files
