+++ src/hoc/withHeadlineLegacyMigration.js
@@ -3,13 +3,13 @@ import { getBlockType } from '@wordpress/blocks';
 import isBlockVersionLessThan from '../utils/check-block-version';
 import wasBlockJustInserted from '../utils/was-block-just-inserted';
 import flexboxAlignment from '../utils/flexbox-alignment';
+import migrateDimensions from './migrations/migrateDimensions';
+import migrateTypography from './migrations/migrateTypography';
+import migrateIconSizing from './migrations/migratingIconSizing';
+import migrateIconPadding from './migrations/migrateIconPadding';
 import {
-	pipe,
+	migrationPipe,
 	updateBlockVersion,
-	migrateIconPadding,
-	migrateIconSizing,
-	migrateTypography,
-	migrateDimensions,
 } from './migrations/utils';
 
 // Set our layout attributes for old Headline blocks.
@@ -54,7 +54,7 @@ export default ( WrappedComponent ) => {
 		const defaults = getBlockType( 'generateblocks/button' )?.attributes;
 
 		useEffect( () => {
-			const newAttributes = pipe(
+			const newAttributes = migrationPipe(
 				attributes,
 				[
 					migrateFlex,
